src/requirements.txt
llama-cpp-python
faiss-cpu
sentence-transformers
PyMuPDF
